+++ spec/rubocop/cop/rails/dynamic_find_by_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe RuboCop::Cop::Rails::DynamicFindBy, :config do
   subject(:cop) { described_class.new(config) }
 
   let(:cop_config) do
-    { 'Whitelist' => %w[find_by_sql] }
+    { 'AllowedMethods' => %w[find_by_sql] }
   end
 
   shared_examples 'register an offense and auto correct' do |message, corrected|
@@ -139,6 +139,32 @@ RSpec.describe RuboCop::Cop::Rails::DynamicFindBy, :config do
     RUBY
   end
 
+  context 'with allowed receiver name' do
+    let(:cop_config) do
+      { 'AllowedReceivers' => %w[Gem::Specification] }
+    end
+
+    it 'accepts dynamic find_by for receiver names in whitelist' do
+      expect_no_offenses(<<~RUBY)
+        Gem::Specification.find_by_name("backend").gem_dir
+      RUBY
+    end
+
+    it 'registers offense for receiver name with no namespace' do
+      expect_offense(<<~RUBY)
+        Specification.find_by_name("backend").gem_dir
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Use `find_by` instead of dynamic `find_by_name`.
+      RUBY
+    end
+
+    it 'registers offense for receiver name with different namespace' do
+      expect_offense(<<~RUBY)
+        RubyGems::Specification.find_by_name("backend").gem_dir
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Use `find_by` instead of dynamic `find_by_name`.
+      RUBY
+    end
+  end
+
   context 'when using safe navigation operator' do
     context 'with dynamic find_by_*' do
       let(:source) { 'user&.find_by_name(name)' }
@@ -148,6 +174,37 @@ RSpec.describe RuboCop::Cop::Rails::DynamicFindBy, :config do
         'Use `find_by` instead of dynamic `find_by_name`.',
         'user&.find_by(name: name)'
       )
+    end
+  end
+
+  # Whitelisted config will be deprecated.
+  context 'with WhiteListed config' do
+    context 'allowed class-names' do
+      let(:cop_config) do
+        { 'Whitelist' => %w[Specification] }
+      end
+
+      # `Whitelist` should not allow receivers for API compatibility.
+      # New API `AllowedReceivers` should be used to add a receiver
+      # to the allowed receivers.
+      it 'registers offense for class methods in whitelist' do
+        expect_offense(<<~RUBY)
+          Specification.find_by_name("backend").gem_dir
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Use `find_by` instead of dynamic `find_by_name`.
+        RUBY
+      end
+    end
+
+    context 'allowed method-names' do
+      let(:cop_config) do
+        { 'Whitelist' => %w[find_by_name] }
+      end
+
+      it 'allows for class methods in whitelist' do
+        expect_no_offenses(<<~RUBY)
+          User.find_by_name("backend").gem_dir
+        RUBY
+      end
     end
   end
 end